+++ src/main/java/com/example/connect/domain/pointuse/repository/PointUseRepository.java
@@ -10,11 +10,11 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public interface PointUseRepository extends JpaRepository<PointUse, Long> {
-    @Query("select pu from PointUse pu where pu.point.id = (select p.id from Point p where p.user.id = :userId and p.isZero = false order by p.createdAt asc limit 1)")
-    PointUse findByOldestPoint(Long userId);
+    @Query("select pu from PointUse pu join pu.point p where p.user.id = :userId and p.isZero = false order by p.createdAt asc")
+    List<PointUse> findByOldestPoint(Long userId, Pageable pageable);
 
-    @Query("select pu from PointUse pu join pu.point p where p.user.id = :userId and pu.pointChange > 0 order by pu.createdAt desc, pu.point.id desc limit 1")
-    PointUse findPointInUse(Long userId);
+    @Query("select pu from PointUse pu join pu.point p where p.user.id = :userId and pu.pointChange > 0 order by pu.createdAt desc, pu.point.id desc")
+    List<PointUse> findPointInUse(Long userId, Pageable pageable);
 
     @Query("select pu from PointUse pu join pu.point p where p.createdAt < :expiredDate and p.isZero = false and pu.createdAt = (select max(ipu.createdAt) from PointUse ipu where ipu.point.id = p.id)")
     List<PointUse> findExpiredPoints(LocalDateTime expiredDate);
